[OVERFLOW-336,344,337] Implement manage members (add, edit, delete)
