+++ backend/app/GraphQL/Mutations/AddMember.php
@@ -15,10 +15,10 @@ final class AddMember
      */
     public function __invoke($_, array $args)
     {
-        $authId = Auth::id();
+        $authUser = Auth::user();
         $team = Team::find($args['team_id']);
 
-        if ($authId != $team->teamLeader->id) {
+        if ($authUser->id != $team->teamLeader->id && $authUser->role_id != 1) {
             throw new CustomException('You are not allowed to add members to this team');
         }
